+++ example/lib/examples/extended_example/radio_example_part.dart
@@ -7,20 +7,13 @@ Widget _buildRadioExample(GroupingType groupingType, {Axis direction}) {
       padding: const EdgeInsets.all(10.0),
       child: Center(
         child: GroupButton(
+          controller: GroupButtonController(selectedIndex: 1),
           direction: direction,
           groupingType: groupingType,
           onSelected: (index, isSelected) =>
               debugPrint('$index button is selected'),
           onDisablePressed: (i) => debugPrint('Disable Button #$i pressed'),
-          buttons: const [
-            "12:00",
-            "13:00",
-            "14:30",
-            "18:00",
-            "19:00",
-            "21:40",
-          ],
-          selectedButton: 5,
+          buttons: const ['12:00', '13:00', '14:30', '18:00', '19:00', '21:40'],
         ),
       ),
     ),