+++ test/test_config.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:group_button/group_button.dart';
 
 class GroupButtonTestWidget extends StatefulWidget {
-  const GroupButtonTestWidget();
+  const GroupButtonTestWidget({Key? key}) : super(key: key);
   @override
   _GroupButtonTestWidgetState createState() => _GroupButtonTestWidgetState();
 }
@@ -15,7 +15,7 @@ class _GroupButtonTestWidgetState extends State<GroupButtonTestWidget> {
     return Scaffold(
       body: GroupButton(
         buttons: buttons,
-        disabledButtons: const [2],
+        controller: GroupButtonController(disabledIndexes: const [2]),
         onSelected: (index, _) {
           setState(() => buttons.removeAt(index));
         },